:art: moves modals/snackbar into a UI folder
This commit takes the Modals and Snackbar fodlers and moves them into the UI folder for improved structure
