+++ client/src/components/UI/Modals/TrainingSeriesModal.js
