+++ client/src/components/UI/Modals/addToTrainingSeriesModal.js
